Switch transport read interface to match that of wandio

+++ lib/bgpstream_transport.c
@@ -89,8 +89,8 @@ bgpstream_transport_t *bgpstream_transport_create(bgpstream_resource_t *res)
   return NULL;
 }
 
-ssize_t bgpstream_transport_read(bgpstream_transport_t *transport,
-                                 uint8_t *buffer, size_t len)
+int64_t bgpstream_transport_read(bgpstream_transport_t *transport,
+                                 uint8_t *buffer, int64_t len)
 {
   return transport->read(transport, buffer, len);
 }
